Flush economy log when bot is shutting down.
